+++ changelog.md
@@ -1,7 +1,8 @@
 # Version 1.1
 
+- Added wool wand
 - Added particle effects to all wands
-- Add ```/wands free``` command
+- Added ```/wands free``` command
 - Made summoners wand less likely to spawn wolfs inside of blocks
 - Increased the price of fireball wand (1 -> 2)